extract bot select modal
