+++ app/public/src/pages/component/avatar.css
@@ -23,8 +23,6 @@
   overflow: hidden;
   white-space: nowrap;
   text-overflow: ellipsis;
-  font-size: 1.5rem;
-  line-height: 1;
 }
 
 .player-title-role {
