+++ app/public/src/pages/component/checkbox/checkbox.css
@@ -1,10 +1,11 @@
 .checkbox-container {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    cursor: var(--cursor-hover);
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  cursor: var(--cursor-hover);
+  margin: 0;
 }
 
 .checkbox-container span {
-    line-height: 16px;
+  line-height: 16px;
 }
