+++ app/public/src/pages/preparation.css
@@ -24,7 +24,7 @@
   flex: 0 1 30em;
 }
 
-.preparation-page footer {
+.preparation-page > footer {
   position: absolute;
   bottom: 0;
   left: 0;
